JENKINS-41281 Move BlueTestResultFactory to blueocean-rest
